requirements.txt
click
python-registry
python-decouple
cryptography
impacket
git+https://github.com/MWedl/pyreadpartitions

